requirements.txt
confluent-kafka>=2.3.0
faker>=19.0.0
fastapi>=0.103.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=15.0.0
pyspark>=3.4.0
requests>=2.31.0
sentence-transformers>=2.2.2
snowflake-connector-python>=3.0.0
uvicorn[standard]>=0.23.0
kafka-python-ng
feedparser
beautifulsoup4
arxiv>=2.0.0
